README.MD
This code is to write a word cloud from a script. We need to remove the punctuations, ignore case and all those words that do not contain alphabets and we have to ignore the uninteresting words. The output will be a dictionary which contains the count of the frequencies of the interesting words