README.md
This code is to write a word cloud from a script. We need to remove the punctuations, ignore the cases and all those words that do not contain alphabets, and ignore the uninteresting words. The output will be a dictionary which includes the count of the frequencies of the exciting words.